Extensions: WPSC: Establish Folder Hierarchy
Note that this has ESLint complain about a lot of class names, but I'll take that over flat folder file soup any day.
