+++ client/extensions/wp-super-cache/app/controller.js
