+++ client/extensions/wp-super-cache/components/advanced/directly-cached-files.jsx
@@ -14,7 +14,7 @@ import FormLabel from 'components/forms/form-label';
 import FormSettingExplanation from 'components/forms/form-setting-explanation';
 import FormTextInput from 'components/forms/form-text-input';
 import SectionHeader from 'components/section-header';
-import WrapSettingsForm from './wrap-settings-form';
+import WrapSettingsForm from '../wrap-settings-form';
 
 class DirectlyCachedFiles extends Component {
 	onKeyDown = event => {
