+++ client/extensions/wp-super-cache/components/advanced/index.jsx
@@ -17,11 +17,11 @@ import ExpiryTime from './expiry-time';
 import FixConfig from './fix-config';
 import LockDown from './lock-down';
 import Miscellaneous from './miscellaneous';
-import QueryStatus from './data/query-status';
+import QueryStatus from '../data/query-status';
 import RejectedUserAgents from './rejected-user-agents';
-import WrapSettingsForm from './wrap-settings-form';
+import WrapSettingsForm from '../wrap-settings-form';
 import { getSelectedSiteId } from 'state/ui/selectors';
-import { getStatus } from './state/status/selectors';
+import { getStatus } from '../../state/status/selectors';
 
 const AdvancedTab = ( {
 	fields: {
