+++ client/extensions/wp-super-cache/components/data/query-settings/index.jsx
@@ -7,8 +7,8 @@ import { connect } from 'react-redux';
 /**
  * Internal dependencies
  */
-import { isRequestingSettings } from '../../state/settings/selectors';
-import { requestSettings } from '../../state/settings/actions';
+import { isRequestingSettings } from '../../../state/settings/selectors';
+import { requestSettings } from '../../../state/settings/actions';
 
 class QuerySettings extends Component {
 	componentWillMount() {
